requirements.txt
discord.py
jsonpickle